requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
graphviz==0.19.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kavenegar==1.1.2
MarkupSafe==2.0.1
model-mommy==2.0.0
openapi-codec==1.3.2
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
